requirements.txt
Django~=5.0
django-crispy-forms~=2.1
crispy-tailwind~=0.5.2
python-decouple~=3.8
bleach~=6.1.0 # For sanitizing reviews
# Production server recommendations
gunicorn~=21.2.0
psycopg2-binary~=2.9.9